refactor: move region to its own file
