+++ cmd/create.go
@@ -556,58 +556,6 @@ var accountCmd = &cobra.Command{
 	},
 }
 
-// createRegionCmd represents the create command
-var createRegionCmd = &cobra.Command{
-	Use:                   "region",
-	Short:                 "setup AppPack resources for an AWS region",
-	Long:                  "*Requires AWS credentials.*",
-	DisableFlagsInUseLine: true,
-	Run: func(cmd *cobra.Command, args []string) {
-		answers, err := askForMissingArgs(cmd, nil)
-		checkErr(err)
-		sess, err := awsSession()
-		checkErr(err)
-		ssmSvc := ssm.New(sess)
-		if createChangeSet {
-			fmt.Println("Creating Cloudformation Change Set for region-level resources...")
-		} else {
-			fmt.Println("Creating region-level resources...")
-		}
-		startSpinner()
-		region := sess.Config.Region
-		tags := []*ssm.Tag{
-			{Key: aws.String("apppack:region"), Value: region},
-			{Key: aws.String("apppack"), Value: aws.String("true")},
-		}
-		_, err = ssmSvc.PutParameter(&ssm.PutParameterInput{
-			Name:  aws.String("/apppack/account/dockerhub-access-token"),
-			Value: getArgValue(cmd, answers, "dockerhub-access-token", true),
-			Type:  aws.String("SecureString"),
-			Tags:  tags,
-		})
-		checkErr(err)
-		cfnTags := []*cloudformation.Tag{
-			{Key: aws.String("apppack:region"), Value: region},
-			{Key: aws.String("apppack"), Value: aws.String("true")},
-		}
-
-		input := cloudformation.CreateStackInput{
-			StackName:   aws.String(fmt.Sprintf("apppack-region-%s", *region)),
-			TemplateURL: aws.String(regionFormationURL),
-			Parameters: []*cloudformation.Parameter{
-				{
-					ParameterKey:   aws.String("DockerhubUsername"),
-					ParameterValue: getArgValue(cmd, answers, "dockerhub-username", true),
-				},
-			},
-			Capabilities: []*string{aws.String("CAPABILITY_IAM")},
-			Tags:         cfnTags,
-		}
-		err = createStackOrChangeSet(sess, &input, createChangeSet, fmt.Sprintf("%s region", *region))
-		checkErr(err)
-	},
-}
-
 // createClusterCmd represents the create command
 var createClusterCmd = &cobra.Command{
 	Use:                   "cluster [<name>]",
@@ -1166,9 +1114,6 @@ func init() {
 	createCmd.PersistentFlags().StringVar(&region, "region", "", "AWS region to create resources in")
 
 	createCmd.AddCommand(accountCmd)
-	createCmd.AddCommand(createRegionCmd)
-	createRegionCmd.Flags().StringP("dockerhub-username", "u", "", "Docker Hub username")
-	createRegionCmd.Flags().StringP("dockerhub-access-token", "t", "", "Docker Hub Access Token (https://hub.docker.com/settings/security)")
 
 	createCmd.AddCommand(appCmd)
 	appCmd.Flags().SortFlags = false
